+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # locate-character changelog
 
+## 2.0.0
+
+* BREAKING: `startIndex` must be passed as a property of `options`, not as a standalone argument
+* Support `offsetLine` and `offsetColumn`
+
 ## 1.0.0
 
 * First release